Refactor: unify confirmation view
