+++ src/components/tx/DecodedTx/index.tsx
@@ -2,10 +2,8 @@ import { type SyntheticEvent, type ReactElement, memo } from 'react'
 import { isCustomTxInfo } from '@/utils/transaction-guards'
 import { Accordion, AccordionDetails, AccordionSummary, Box, Skeleton, Stack } from '@mui/material'
 import { OperationType, type SafeTransaction } from '@safe-global/safe-core-sdk-types'
-import type { DecodedDataResponse } from '@safe-global/safe-gateway-typescript-sdk'
+import type { DecodedDataResponse, TransactionDetails } from '@safe-global/safe-gateway-typescript-sdk'
 import { Operation } from '@safe-global/safe-gateway-typescript-sdk'
-import useChainId from '@/hooks/useChainId'
-import ErrorMessage from '../ErrorMessage'
 import Summary, { PartialSummary } from '@/components/transactions/TxDetails/Summary'
 import { trackEvent, MODALS_EVENTS } from '@/services/analytics'
 import Multisend from '@/components/transactions/TxDetails/TxData/DecodedData/Multisend'
@@ -13,13 +11,10 @@ import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
 import DecodedData from '@/components/transactions/TxDetails/TxData/DecodedData'
 import accordionCss from '@/styles/accordion.module.css'
 import HelpToolTip from './HelpTooltip'
-import { useGetTransactionDetailsQuery } from '@/store/gateway'
-import { skipToken } from '@reduxjs/toolkit/query/react'
-import { asError } from '@/services/exceptions/utils'
 
 type DecodedTxProps = {
   tx?: SafeTransaction
-  txId?: string
+  txDetails: TransactionDetails
   showMultisend?: boolean
   decodedData?: DecodedDataResponse
   showMethodCall?: boolean
@@ -36,40 +31,26 @@ export const Divider = () => (
 
 const DecodedTx = ({
   tx,
-  txId,
+  txDetails,
   decodedData,
   showMultisend = true,
   showMethodCall = false,
 }: DecodedTxProps): ReactElement => {
-  const chainId = useChainId()
   const isMultisend = !!decodedData?.parameters?.[0]?.valueDecoded
   const isMethodCallInAdvanced = !showMethodCall || (isMultisend && showMultisend)
-
-  const {
-    data: txDetails,
-    error: txDetailsError,
-    isLoading: txDetailsLoading,
-  } = useGetTransactionDetailsQuery(
-    chainId && txId
-      ? {
-          chainId,
-          txId,
-        }
-      : skipToken,
-  )
 
   const onChangeExpand = (_: SyntheticEvent, expanded: boolean) => {
     trackEvent({ ...MODALS_EVENTS.TX_DETAILS, label: expanded ? 'Open' : 'Close' })
   }
 
-  const addressInfoIndex = txDetails?.txData?.addressInfoIndex
+  const addressInfoIndex = txDetails.txData?.addressInfoIndex
 
   const txData = {
     dataDecoded: decodedData,
     to: { value: tx?.data.to || '' },
     value: tx?.data.value,
     operation: tx?.data.operation === OperationType.DelegateCall ? Operation.DELEGATE : Operation.CALL,
-    trustedDelegateCallTarget: txDetails?.txData?.trustedDelegateCallTarget ?? true,
+    trustedDelegateCallTarget: txDetails.txData?.trustedDelegateCallTarget ?? true,
     addressInfoIndex,
   }
 
@@ -90,7 +71,7 @@ const DecodedTx = ({
         </Box>
       )}
 
-      {isMultisend && showMultisend && <Multisend txData={txDetails?.txData || txData} compact />}
+      {isMultisend && showMultisend && <Multisend txData={txDetails.txData || txData} compact />}
 
       <Box>
         <Accordion elevation={0} onChange={onChangeExpand} sx={!tx ? { pointerEvents: 'none' } : undefined}>
@@ -113,14 +94,7 @@ const DecodedTx = ({
                 <Divider />
               </>
             )}
-
-            {txDetails ? <Summary txDetails={txDetails} defaultExpanded /> : tx && <PartialSummary safeTx={tx} />}
-
-            {txDetailsLoading && <Skeleton />}
-
-            {txDetailsError && (
-              <ErrorMessage error={asError(txDetailsError)}>Failed loading all transaction details</ErrorMessage>
-            )}
+            <Summary txDetails={txDetails} defaultExpanded /> : tx && <PartialSummary safeTx={tx} />
           </AccordionDetails>
         </Accordion>
       </Box>
