+++ src/components/tx/confirmation-views/types.d.ts
@@ -1,0 +1,6 @@
+import type { TransactionDetails } from '@safe-global/safe-gateway-typescript-sdk'
+
+export type NarrowConfirmationViewProps = {
+  txDetails: TransactionDetails
+  txInfo: TransactionDetails['txInfo']
+}
